refactor(forside): flytt hooks og utils til utils-katalog
Samlar bestemSøknadsstart-logikken (forsideUtils), useStartSøknad og
useFjernPlanleggerDataFraUrl i ein ny pages/forside/utils-katalog, og
oppdaterer importane i Forside.tsx og ForeldrepengesøknadRoutes.tsx.
